feat: include github templates
Having templates for PRs helps maintain document the intentional
changes that a PR introduces. This helps a lot with debugging
features in the future, and for those who did not author the
feature to determine a good remediation to related issues.

Issue templates help us by re-directing the issues to be opened
against the main repository